requirements.txt
openai
python-dotenv
numpy
pandas
pymysql
requests
beautifulsoup4
python-dateutil
tiktoken
lxml
matplotlib
seaborn
ipython
jsonpath 